fix: set background color style more globally
This is ugly, there is likely a more specific place this should go. It is also set on `pre`, but the inline style that the markdown processor sets on one of its containing divs needs the variable to be set already (and targeting that specific div seems potentially more likely to break).