chore: re-add ical link feature and rename “singupLink” 🎶
this adds the ical column back to Design Systems Week tables, and it renames singupLink (!) to sessionLink

It also adds all ical links from 2013 so that that column doesn't show up empty. They aren't relevant, because they are invites for something in the past, but also not broken, so this seems neatest.

And it removes sessionlinks from 2023 as they no longer exist on the gebruikercentraal website

Signed-off-by: Hidde de Vries <hidde@hiddedevries.nl>