chore: reference correct package in changesets
All these changes are for documentation repo, not nlds-design-tokens. The default is also documentation repo, which is why we updated the README.md in the .changeset folder.

Found when reviewing a PR with Renate.

Next step is to merge to release package, which was created in oct 2024: https://github.com/nl-design-system/documentatie/pull/1540.